Refactored namespaces for better clarity of the code
