+++ client.wpf/OneTrueError.Client.Wpf/Presenters/NotificationControlPresenter.cs
@@ -1,4 +1,4 @@
-﻿namespace OneTrueError.Client.Wpf
+﻿namespace OneTrueError.Client.Wpf.Presenters
 {
     public class NotificationControlPresenter
     {
