+++ client.wpf/OneTrueError.Client.Wpf/Presenters/UserErrorDescriptionPresenter.cs
@@ -1,4 +1,4 @@
-﻿namespace OneTrueError.Client.Wpf
+﻿namespace OneTrueError.Client.Wpf.Presenters
 {
     public class UserErrorDescriptionPresenter
     {
