+++ client.wpf/OneTrueError.Client.Wpf/Views/ErrorMessageView.xaml
@@ -1,0 +1,27 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:presenters="clr-namespace:OneTrueError.Client.Wpf.Presenters">
+
+    <DataTemplate DataType="{x:Type presenters:ErrorMessagePresenter}">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="2.5*"/>
+            </Grid.RowDefinitions>
+
+            <TextBlock Grid.Row="0"
+                       Text="Error Message"
+                       FontWeight="Bold"
+                       FontSize="14"
+                       />
+            <TextBlock Grid.Row="1"
+                       Text="{Binding ExceptionMessage}"
+                       FontWeight="Bold"
+                       Foreground="#33A8F7"
+                       Margin="0,5,0,5"
+                       />
+        </Grid>
+    </DataTemplate>
+
+
+</ResourceDictionary>
