+++ client.wpf/OneTrueError.Client.Wpf/Views/NotificationControlView.xaml
@@ -1,0 +1,24 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:presenters="clr-namespace:OneTrueError.Client.Wpf.Presenters">
+    <DataTemplate DataType="{x:Type presenters:NotificationControlPresenter}">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="*"/>
+            </Grid.RowDefinitions>
+
+            <TextBlock Grid.Row="0"
+                       Text="Status updates"
+                       FontWeight="Bold"
+                       FontSize="14"
+                       />
+            <TextBlock Grid.Row="1"
+                       Text="Enter your email address if you would like to receive status updates about this error."
+                       />
+            <TextBox Grid.Row="2"
+                     Text="{Binding Email}"/>
+        </Grid>
+    </DataTemplate>
+</ResourceDictionary>
