+++ client.wpf/OneTrueError.Client.Wpf/Views/ReportDialogView.xaml
@@ -1,11 +1,19 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:OneTrueError.Client.Wpf">
+                    xmlns:local="clr-namespace:OneTrueError.Client.Wpf"
+                    xmlns:converters="clr-namespace:OneTrueError.Client.Wpf.Converters"
+                    xmlns:presenters="clr-namespace:OneTrueError.Client.Wpf.Presenters">
+    
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="ErrorMessageView.xaml"/>
+        <ResourceDictionary Source="UserErrorDescriptionView.xaml"/>
+        <ResourceDictionary Source="NotificationControlView.xaml"/>
+    </ResourceDictionary.MergedDictionaries>
     
     <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter"/>
-    <local:BooleanToGridLengthConverter x:Key="BooleanToGridLengthConverter"/>
+    <converters:BooleanToGridLengthConverter x:Key="BooleanToGridLengthConverter"/>
     
-    <DataTemplate DataType="{x:Type local:ErrorReportDialogPresenter}">
+    <DataTemplate DataType="{x:Type presenters:ErrorReportDialogPresenter}">
         <Grid>
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="200" />
@@ -24,7 +32,7 @@
                    Grid.RowSpan="4"
                    Stretch="UniformToFill"
                    StretchDirection="Both"
-                   Source="Resources/bigstock-Output.jpg"
+                   Source="../Resources/bigstock-Output.jpg"
                    Margin="1"
                />
 
@@ -71,67 +79,6 @@
         </Grid>
     </DataTemplate>
 
-    <DataTemplate DataType="{x:Type local:ErrorMessagePresenter}">
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="2.5*"/>
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0"
-                       Text="Error Message"
-                       FontWeight="Bold"
-                       FontSize="14"
-                       />
-            <TextBlock Grid.Row="1"
-                       Text="{Binding ExceptionMessage}"
-                       FontWeight="Bold"
-                       Foreground="#33A8F7"
-                       Margin="0,5,0,5"
-                       />
-        </Grid>
-    </DataTemplate>
-
-    <DataTemplate DataType="{x:Type local:UserErrorDescriptionPresenter}">
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="7*"/>
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0"
-                       Text="Feedback"
-                       FontWeight="Bold"
-                       FontSize="14"
-                       />
-            <TextBlock Grid.Row="1"
-                       Text="Could you please let us know how to reproduce it?"
-                       />
-            <TextBox Grid.Row="2"
-                     Text="{Binding UserDescription}"/>
-        </Grid>
-    </DataTemplate>
-
-    <DataTemplate DataType="{x:Type local:NotificationControlPresenter}">
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0"
-                       Text="Status updates"
-                       FontWeight="Bold"
-                       FontSize="14"
-                       />
-            <TextBlock Grid.Row="1"
-                       Text="Enter your email address if you would like to receive status updates about this error."
-                       />
-            <TextBox Grid.Row="2"
-                     Text="{Binding Email}"/>
-        </Grid>
-    </DataTemplate>
+    
 
 </ResourceDictionary>
