+++ client.wpf/OneTrueError.Client.Wpf/Views/UserErrorDescriptionView.xaml
@@ -1,0 +1,25 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:wpf="clr-namespace:OneTrueError.Client.Wpf.Presenters">
+    <DataTemplate DataType="{x:Type wpf:UserErrorDescriptionPresenter}">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="7*"/>
+            </Grid.RowDefinitions>
+
+            <TextBlock Grid.Row="0"
+                       Text="Feedback"
+                       FontWeight="Bold"
+                       FontSize="14"
+                       />
+            <TextBlock Grid.Row="1"
+                       Text="Could you please let us know how to reproduce it?"
+                       />
+            <TextBox Grid.Row="2"
+                     Text="{Binding UserDescription}"/>
+        </Grid>
+    </DataTemplate>
+
+</ResourceDictionary>
